Move shared method to a helper
Codeclimate was complaining about this as a duplicity, so I moved it to
a helper. The helper itself doesn't have tests, because it's covered in
the respective controllers.
