+++ app/controllers/deleted_applications_controller.rb
@@ -1,6 +1,8 @@
 class DeletedApplicationsController < ApplicationController
   before_action :authenticate_user!
 
+  include ProcessedViewsHelper
+
   def index
     @applications = Query::DeletedApplications.new(current_user).find.map do |application|
       Views::ApplicationList.new(application)
@@ -15,12 +17,5 @@ class DeletedApplicationsController < ApplicationController
 
   def application
     @application ||= Application.find(params[:id])
-  end
-
-  def assign_views
-    @application = application
-    @processed = Views::ProcessingDetails.new(application)
-    @overview = Views::ApplicationOverview.new(application)
-    @result = Views::ApplicationResult.new(application)
   end
 end
