+++ app/controllers/processed_applications_controller.rb
@@ -1,6 +1,8 @@
 class ProcessedApplicationsController < ApplicationController
   before_action :authenticate_user!
 
+  include ProcessedViewsHelper
+
   def index
     @applications = Query::ProcessedApplications.new(current_user).find.map do |application|
       Views::ApplicationList.new(application)
@@ -29,13 +31,6 @@ class ProcessedApplicationsController < ApplicationController
 
   def application
     @application ||= Application.find(params[:id])
-  end
-
-  def assign_views
-    @application = application
-    @processed = Views::ProcessingDetails.new(application)
-    @overview = Views::ApplicationOverview.new(application)
-    @result = Views::ApplicationResult.new(application)
   end
 
   def delete_params
