+++ app/helpers/processed_views_helper.rb
@@ -1,0 +1,8 @@
+module ProcessedViewsHelper
+  def assign_views
+    @application = application
+    @processed = Views::ProcessingDetails.new(application)
+    @overview = Views::ApplicationOverview.new(application)
+    @result = Views::ApplicationResult.new(application)
+  end
+end
